pyproject.toml
[project]
name = "job-recommendation-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp==3.9.1",
    "alembic==1.13.1",
    "asyncpg==0.29.0",
    "email-validator==2.1.0",
    "fastapi==0.109.0",
    "httpx==0.26.0",
    "numpy==1.26.3",
    "passlib[bcrypt]==1.7.4",
    "pgvector==0.2.4",
    "psycopg2-binary==2.9.9",
    "pydantic==2.5.3",
    "pydantic-settings==2.1.0",
    "python-dateutil==2.8.2",
    "python-dotenv==1.0.0",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.6",
    "sentence-transformers==2.3.1",
    "sqlalchemy==2.0.25",
    "torch==2.2.0",
    "uvicorn[standard]==0.27.0",
]
